using float instead of double

+++ dumb_calculator.pas
@@ -6,8 +6,8 @@ uses
 getopts, SysUtils;
 
 var 
-operant_1 : Double;
-operant_2 : Double;
+operant_1 : Float;
+operant_2 : Float;
 options : array of TOption = (
     (Name: 'add';
     has_arg:  0;
@@ -16,7 +16,7 @@ options : array of TOption = (
     );
 c : char = #0;
 option_index : LongInt;
-math_operation: function(op1, op2 : Double) : Double;
+math_operation: function(op1, op2 : Float) : Float;
 
 procedure wrong_arguments_number();
 begin
@@ -25,7 +25,7 @@ begin
     ExitCode := 1;
 end;
 
-function addition(op1, op2 : Double) : Double;
+function addition(op1, op2 : Float) : Float;
 begin
     addition := op1 + op2
 end;
